requirements.txt
Keras==1.0.8
numpy==1.11.1
Pillow==3.3.1
PyYAML==3.12
scikit-learn==0.17.1
scipy==0.18.0
six==1.10.0
Theano==0.8.2
